Crate srml_council

Council system: Handles the voting in and maintenance of council members.

Re-exports

pub use crate::seats::Trait;

pub use crate::seats::Module;

pub use crate::seats::RawEvent;

pub use crate::seats::Event;

pub use crate::seats::VoteIndex;

Modules

motions
Council voting system.

seats
Council system: Handles the voting in and maintenance of council members.

Traits

OnMembersChanged
Trait for type that can handle incremental changes to a set of account IDs.