Module srml_council::seats

Council system: Handles the voting in and maintenance of council members.

Structs

ActiveCouncil
The current council. When there's a vote going on, this should still be used for executive matters. The block number (second element in the tuple) is the block that their position is active until (calculated by the sum of the block number when the council member was elected and their term duration).

ApprovalsOf
A list of votes for each voter. The votes are stored as numeric values and parsed in a bit-wise manner.

CandidacyBond
How much should be locked up in order to submit one's candidacy.

CandidateCount
Current number of active candidates

Candidates
The present candidate list.

CarryCount
How many runners-up should have their approvals persist until the next vote.

DecayRatio
Decay factor of weight when being accumulated. It should typically be set to at least council_size -1 to keep the council secure. When set to N, it indicates (1/N)^t of staked is decayed at weight increment step t. 0 will result in no weight being added at all (normal approval voting).

DesiredSeats
Number of accounts that should be sitting on the council.

GenesisConfig

InactiveGracePeriod
How many vote indices need to go by after a target voter's last vote before they can be reaped if their approvals are moot.

Leaderboard
Get the leaderboard if we're in the presentation phase. The first element is the weight of each entry; It may be the direct summed approval stakes, or a weighted version of it.

Module

NextFinalize
The accounts holding the seats that will become free on the next tally.

NextVoterSet
the next free set to store a voter in. This will keep growing.

PresentSlashPerVoter
The punishment, per voter, if you provide an invalid presentation.

PresentationDuration
How long to give each top candidate to present themselves after the vote ends.

RegisterInfoOf
The vote index and list slot that the candidate who was registered or None if they are not currently registered.

TermDuration
How long each position is active for.

VoteCount
The total number of vote rounds that have happened or are in progress.

VoterCount
Current number of Voters.

VoterInfo
The activity status of a voter.

VoterInfoOf
Basic information about a voter.

Voters
The present voter list (chunked and capped at VOTER_SET_SIZE).

VotingBond
How much should be locked up in order to be able to submit votes.

VotingFee
The amount of fee paid upon each vote submission, unless if they submit a hole index and replace it.

VotingPeriod
How often (in blocks) to check for new votes.

Enums

Call

CellStatus
Used to demonstrate the status of a particular index in the global voter list.

RawEvent
Events for this module.

Constants

APPROVAL_FLAG_LEN

APPROVAL_FLAG_MASK

APPROVAL_SET_SIZE

VOTER_SET_SIZE

Traits

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait

VoteIndex