Struct srml_council::seats::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: 'static + Trait> Module<T>

pub fn candidacy_bond(
) -> <<T as Trait>::Currency as Currency<<T as Trait>::AccountId>>::Balance

Amount that must be locked up in order to submit one's candidacy.

pub fn voting_bond(
) -> <<T as Trait>::Currency as Currency<<T as Trait>::AccountId>>::Balance

Amount that must be locked up in order to be able to submit votes.

pub fn present_slash_per_voter(
) -> <<T as Trait>::Currency as Currency<<T as Trait>::AccountId>>::Balance

The punishment, per voter, if you provide an invalid presentation.

pub fn carry_count() -> u32

Number of runners-up who should have their approvals persist until the next vote.

pub fn presentation_duration() -> T::BlockNumber

Number of blocks to give each top candidate to present themselves after the vote ends.

pub fn inactivity_grace_period() -> VoteIndex

Number of vote indices that need to go by after a target voter's last vote before they can be reaped if their approvals are moot.

pub fn voting_period() -> T::BlockNumber

How often (in blocks) to check for new votes.

pub fn term_duration() -> T::BlockNumber

How long (in blocks) each position is active for.

pub fn desired_seats() -> u32

Number of accounts that should be sitting on the council.

pub fn active_council() -> Vec<(T::AccountId, T::BlockNumber)>

The current council. When there's a vote going on, this should still be used for executive matters. The block number is the block that the associated account ID's position is active until (calculated by the sum of the block number when the council member was elected and their term duration).

pub fn vote_index() -> VoteIndex

The total number of votes that have happened or are in progress.

pub fn approvals_of<K: Borrow<T::AccountId>>(key: K) -> Vec<bool>

A list of votes for each voter, respecting the last cleared vote index that this voter was last active at.

pub fn candidate_reg_info<K: Borrow<T::AccountId>>(
key: K
) -> Option<(VoteIndex, u32)>

The vote index and list slot in which the candidate account ID was registered or None if they are not currently registered.

pub fn voter_last_active<K: Borrow<T::AccountId>>(key: K) -> Option<VoteIndex>

The last cleared vote index that this voter was last active at.

pub fn voters() -> Vec<T::AccountId>

The present voter list.

pub fn candidates() -> Vec<T::AccountId>

The present candidate list.

pub fn candidate_count() -> u32

Number of candidates.

pub fn next_finalize() -> Option<(T::BlockNumber, u32, Vec<T::AccountId>)>

The accounts holding the seats that will become free on the next tally. Tuple of the block number at which the next tally will occur, the number of seats, and a list of the account IDs.

pub fn snapshoted_stakes(
) -> Vec<<<T as Trait>::Currency as Currency<<T as Trait>::AccountId>>::Balance>

The stakes as they were at the point that the vote ended.

pub fn leaderboard(
) -> Option<Vec<(<<T as Trait>::Currency as Currency<<T as Trait>::AccountId>>::Balance, T::AccountId)>>

Get the leaderboard if we're in the presentation phase.

impl<T: Trait> Module<T>

pub fn presentation_active() -> bool

Returns true if we're currently in a presentation period.

pub fn is_a_candidate(who: &T::AccountId) -> bool

Returns true if who is a candidate at the moment.

pub fn next_vote_from(n: T::BlockNumber) -> T::BlockNumber

Determine the block that a vote can happen on, which is no less than n.

pub fn next_tally() -> Option<T::BlockNumber>

The block number on which the tally for the next election will happen. None only if the desired seats of the council is zero.