Trait srml_council::OnMembersChanged

pub trait OnMembersChanged<AccountId> {
    fn on_members_changed(new: &[AccountId], old: &[AccountId]);
}

Trait for type that can handle incremental changes to a set of account IDs.

Required methods

fn on_members_changed(new: &[AccountId], old: &[AccountId])

A number of members new just joined the set and replaced some old ones.

Loading content...