Module srml_council::voting

Council proposal voting, tallying, and vetoing.

Structs

CooloffPeriod
Period (in blocks) that a veto is in effect.

CouncilVoteOf
Map from a proposal ID and voter to the outcome of the vote. True indicates that it passed.

EnactDelayPeriod
Number of blocks by which to delay enactment of successful, non-unanimous, council-instigated referendum proposals.

GenesisConfig

Module

ProposalOf
Map from proposal ID to the proposal.

ProposalVoters
List of voters that have voted on a proposal ID.

Proposals
A list of proposals by block number and proposal ID.

VetoedProposal
A veto of a proposal. The veto has an expiry (block number) and a list of vetoers.

VotingPeriod
Period (in blocks) that a vote is open for.

Enums

Call

RawEvent
Events for this module.

Traits

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait