Module srml_council::seats

Council member seat candidacy, voting, approval, and maintenance.

Structs

ActiveCouncil
The current council. When there's a vote going on, this should still be used for executive matters. The block number is the block that the associated account ID's position is active until (calculated by the sum of the block number when the council member was elected and their term duration).

ApprovalsOf
A list of votes for each voter, respecting the last cleared vote index that this voter was last active at.

CandidacyBond
Amount that must be locked up in order to submit one's candidacy.

CandidateCount
Number of candidates.

Candidates
The present candidate list.

CarryCount
Number of runners-up who should have their approvals persist until the next vote.

DesiredSeats
Number of accounts that should be sitting on the council.

GenesisConfig

InactiveGracePeriod
Number of vote indices that need to go by after a target voter's last vote before they can be reaped if their approvals are moot.

LastActiveOf
The last cleared vote index that this voter was last active at.

Leaderboard
Get the leaderboard if we're in the presentation phase.

Module

NextFinalize
The accounts holding the seats that will become free on the next tally. Tuple of the block number at which the next tally will occur, the number of seats, and a list of the account IDs.

PresentSlashPerVoter
The punishment, per voter, if you provide an invalid presentation.

PresentationDuration
Number of blocks to give each top candidate to present themselves after the vote ends.

RegisterInfoOf
The vote index and list slot in which the candidate account ID was registered or None if they are not currently registered.

SnapshotedStakes
The stakes as they were at the point that the vote ended.

TermDuration
How long (in blocks) each position is active for.

VoteCount
The total number of votes that have happened or are in progress.

Voters
The present voter list.

VotingBond
Amount that must be locked up in order to be able to submit votes.

VotingPeriod
How often (in blocks) to check for new votes.

Enums

Call

RawEvent
Events for this module.

Traits

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait

VoteIndex
Index for the total number of vote tallies that have happened or are in progress.